div function

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -2,5 +2,5 @@
 -- the first by the second number or returns 0 if the second number is equal to 0
 
 CREATE FUNCTION SafeDiv (a INT, b INT)
-RETURNS INT DETERMINISTIC
-RETURN IF(b = 0, 0, a / b);
+RETURNS FLOAT DETERMINISTIC
+RETURN IF(b = 0, 0.00, a / b);
